setting up for clarity2
