+++ parser/pom.xml
@@ -36,11 +36,11 @@
     </repositories>
 
     <dependencies>
-        <dependency>
-            <groupId>com.skadistats</groupId>
-            <artifactId>clarity</artifactId>
-            <version>1.1</version>
-        </dependency>
+<dependency>
+    <groupId>com.skadistats</groupId>
+    <artifactId>clarity</artifactId>
+    <version>2.0-SNAPSHOT</version>
+</dependency>
         <dependency>
             <groupId>com.google.code.gson</groupId>
             <artifactId>gson</artifactId>
